Remove NetworkConstantsInterface
The purpose of this interface is to enable mocking later; however,
we may have a different approach we want to take, so rolling back
this design for now. It may end up going back in if the other
approach doesn't work out, but until we're sure, we'll play it safe
and avoiding adding extra classes for now.

Change-Id: Id7d742f3f8fb5821abc19ac104315b66d32edbcb