Update annotations extractor to allow merging compat annotations
Normally the annotations extractor will filter out all annotations
for APIs that are not in the API whitelist. However, if we're
merging in annotations XML files that have been explicitly marked
with an API range attribute (stating that the annotation may only
apply for a subset of API levels), those should be merged regardless
of the current API whitelist.

Change-Id: I849cb793cd0322ff4dc0405f4f4f6eb81beb75ae