Get fragment lifecycle state using bytecode manipulation.
In this cl, the FragmentAdapter is created to go through the android
system's fragment class, and do the instrumentation on the state we
interested in. The FragmentWrapper class will provide the method we want
to instrument using INVOKESTATIC.

The method to process onCreateView state corretly could be hacky, (Since
the user may not call super.onCreateView() in his/her override
function). So I decide to put it in another cl.

A wrapper function is added and more detailed documents.

Remove unnecessary TODO in source code.

Change-Id: I47e0838b4c083efeb39f9c7b80f274cfa24e5658