Adds proper groovy support to the bazel build
We used to call the groovyc command line tool to compile
the groovy files along side all the java files. This was done
by calling the system's java instead of bazels. Not good.

This change does the following things:
. It uses the groovy-all prebuilt instead of a custom command line one
. It uses the tools directly from java to simplify shell manipulation.
. It uses two tools out of the groovy suite: genstubs and groovyc:

The groovy files are first compiled to .java stubs.
These stubs are given to the standard java_library rule as
sourcepath arguments. The sources are given as a jar and output
of a genrule. The java .jar is then compiled without these classes
in. Then groovy is called using the newly created java .jar and the
remaining groovy classes are compiled.

Change-Id: Ic938cf0b3b1badbb7eee8147bad78d0860f368e3