Collect all Archive static variables into a single class
In order to run SDK Downloading logic in a reentrant way (download
components for one OS, then stop and download components for another),
we need to reset all the cached static variables.

This is easier done if all static values are wrapped up in a single
class. Resetting them is now done by overwriting the static instance.

Change-Id: I27c21442b6b0b601b69047810ff8ef2eb8b57c50