Adds form compilation
Adds support to instrument .form classes.
This works by processing the .jar after it's been compiled. There is
a concept of nested forms which means that a form can reference another
one potentially in a different module. That's the reason why .form files
are kept as resources inside the original jars.

Change-Id: Ie1abdbaafc668eb9a622bf3651289998d7526e22