b/30948706:  Crash running linter with 2.2-beta1 plugin
This CL fixes a bug in the code which initializes modifier lists.
This was triggered by a particular source construct in the tycho
codebase.

While reproducing this I also noticed some non-fatal warnings
where lint aborts analysis when it can't resolve a symbol (e.g.
where a compiled .class file references a class that isn't
found on lint's class path). This CL handles this a bit more
gracefully.

Change-Id: I90befafb712c2d5bbde52d4bb4dd033180cf43aa