Validate analytics event parameter names
This updates the Analytics event detector to report an error when it
finds an invalid Analytics parameter name. This detector traces the
reference to the Bundle object passed into FirebaseAnalytics#logEvent
(possibly across the file boundary), and tracks all of the calls to
Bundle#putString, Bundle#putLong, and Bundle#putDouble. If one of those
calls uses an invalid value for the first argument, the detector reports
an error at the logEvent call, with a secondary location at the put*
call. The logic isn't complete, but it should cover the common cases and
help developers that are new to the Analytics API.

Documentation: https://firebase.google.com/docs/reference/android/com/google/firebase/analytics/FirebaseAnalytics#logEvent(java.lang.String,%20android.os.Bundle)

Change-Id: I69d3423d6ea56d1b868f2f3925b8a4b01fbdc42a