Handle exception when creating or using the build cache.
When FileCache.createFile(outputFile, inputs, fileProducer) throws an
exception, we can now check whether the exception was caused by
fileProducer or by the cache itself, and handle the case accordingly.

Previously, if a cached file does not exist, we create the cached file
first, then copy it to the final output file. If an exception occurs in
the first step, we do not know whether it is caused by the file producer
or the cache since the file producer is asked to create a file at a
different location than was intended. To address that, we now create the
final output file first, then copy it to the cache. This allows us to
isolate any exceptions thrown by the file producer.

If the cached file does exist, then the implementation is unchanged: we
simply copy the cached file to the output file.

Change-Id: I48aa3b7bc7f025fb137f65d9e792ce81c1b0573b