Ship data binding info in aar
This CL adds a new data-binding folder into the aar files which includes
the artifact information that is only available while compiling the
library. Application compilation later uses this data to generate final
classes etc.

It also removes the existing DataBindingTransform that was only used in
Jack builds. The new transform is designed to be used by both javac and
jack and takes care of exporting data binding artifacts from dependencies.

Bug: 31033895
Change-Id: I55bd685ede98760e0af9cf4fd0397cb9f5031fae