moved the manifest changes to a proper task.
In instant run mode, we check if the textual or binary representation
of the merged manifest file has changed to provoke a full APK build.
This used to be part of the aapt invocation task.

Unfortunately, this can create situations where the manifest changes
are not detected. The aapt task can be invoked during an IDE sync
which is not done in InstantRun mode. Next instant run enabled build
request from the IDE would not be able to see the changed manifest
since the aapt task would not run again.

Also added a safeguard in case the backup manifest/crc files are
not found. In case those backup files are missing, we now force a
full APK.

moved the check to a proper task that's only configured in instant
run mode and will run all the time. That task will compare the latest
manifest file to the one that was last seen during an instant run
build.

Also added some traces.

https://code.google.com/p/android/issues/detail?id=220425

Change-Id: Ie8c65d189dc203b1058866ad62e85657ac0eedeb