Add FileSystemNotifier to monitor files events.
Enabled perfd to process a file only after it has been fully populated
and closed. It is useful in the case of ART traces which are generated
via a call to am with "start" and "stop": The latter is non-blocking
which means am will returns even if ART has not finished writing the
content of the traces file. Using the FileSystemNotifier, perfd can
wait on a "close" event, making sure the full trace content is on disk.
Change-Id: Ia510858571f0626e339d29c082bd39e846cc8d31