Cache the exploded outputs of aar files.
Cache the exploded outputs of (external) aar files using the new build
cache. With the build cache enabled, all exploded aar files for
different projects will be stored at a central location inside the build
cache directory. If the build cache is disabled, these files are still
stored in the project's build/intermediates directory as before.

Note that to avoid potential issues with Gradle incremental builds when
multiple tasks share the same output directory, we do not register the
exploded directory as an output directory when the build cache is
enabled. A small side effect of this is that Gradle will not show an
UP-TO-DATE status even though a task may run very quickly (essentially
equivalent to the task being skipped) when the cache is hit.

Test: Covered by the new PrepareLibraryTaskTest
Change-Id: Iae4ad872a8e50cf560509dfc2f84998930aed6c7