Send CPU trace data in bits instead of an on-device path.
This CL updates the CPU Profiler API to send the bits of trace data
(produced by ART or simpleperf) through gRPC, instead of an on-device
path. The purpose is to simplify the API so it is possible for
Data Store and Perfd to have the same interface.

This CL should be merged together with the known user of the API
(Change ID: I77793b91e1e10148484105ec0b3ce82194bc1d33).

Test: Manual test, during and after the development.
1. Inside a debugger, when saving the bits into a file, the file is
   identical to the one that was transferred by adb.
2. Manually tested new perfd and Studio. CPU profiler's method trace
   works as before when selecting ART sampling based tracing.

Change-Id: I51289970cb2a6963555ede2beceda9fde9f097f5