80680: Make the lint report background more readable
This CL adds a new HTML reporter to Lint which writes
the report in a Material style (using the MDL library)
to modernize the format (the old format was heavily
Holo influenced, using Holo colors and fonts etc, and
feels dated at this point.

Among other improves it also starts syntax highlighting
code snippets.

The old reporter is still available via a flag until
any issues in the new report are ironed out.

Test: New unit tests included

Change-Id: I55fd30ec12b31bb86473121ff02ae3814ac6ef94