Do not emit error messages when class is targeting newer platform.
In InstantRun mode, we now instrument classes against the target API
level the APK will be deployed on. Sometimes, users may have code
that targets newer API level and in case that code subclassed a
class/interface that got added recently in the platform,
we would emit a warning message the we cannot find the parent classs.

Although the warning was innocuous, it's confusing for the user.
With the CL, if we encounter a class that is annotated with
@TargetAPI(x) we will NOT emit the warning if x > targetApiLevel
for the device. Otherwise we will continue emit the warning.

The AsmUtils class was mainly created from code extracted from
IncrementalVisitor for better testability. More unit tests will
be added in a subsequent CL to complete unit testing of this
class.

Bug: 212672
Test: adding unit tests and manually tested with a sample project.

Change-Id: I34b1be670bd8e76e589ea4823d20313242a8037f