Enforce the clients of FileCache to provide a command as input.
Since multiple tasks could use the same cache, we need a way to separate
the cached results of a task from those of other tasks. To do that, we
now enforce the clients of FileCache to provide a command that
identifies the file creator callback function (which usually corresponds
to a Gradle task). The command will be part of the inputs when the
client calls API methods on FileCache.

Test: Covered by new tests in FileCacheTest
Change-Id: I7a06963b4e6ca15866406fa2930b12dbc3d249ab