Move payload file caching logic to NetworkCache
Up until now, the network request body caching logic has all been
contained in the InternalNetworkService class, but in a followup CL,
the file will need to be accessed from NetworkService as well. While
the refactoring done here here is essentially a no-op, moving logic
out of InternalNetworkService into NetworkCache, it will make the
next CL easier to implement.

Test: Although this class doesn't introduce any new logic, NetworkCache
is missing test coverage and will need to be handled in a followup CL.
Added TODO for now.

Change-Id: I718547fdad912a9056d16a6a5c971879603a9603