[AVD Mgr] Don't create userdata.img if it already exists
When renaming an AVD, don't create a new copy of userdata.img
from the SDK. Just use the already-copied version in the AVD
directory.

Note: This problem was a regression caused by switching from
custom copy code to FileOps. The FileOps code throws an exception
if the destination already exists. The custom code silently
overwrote the destination.

Test: Added testRenameAvd() to AvdManagerTest.java

BUG: 227279

Change-Id: Ic3691db62959675f4d8a88618615e926fe24e3cc