Change and rename LibraryGraph.
It is now DependencyGraphs.

Before both graph for compile and package, each their own LibraryGraph
instance, would contain a list of provided/skipped but it doesn't make
sense to have them in both places. In fact provided is only useful for
compile and skipped for package.

So now there's a single DependencyGraphs instance that contains both
the compile and package graphs as well as the provided/skipped
information.

We have 2 implementations, one for the default mode and one for the
full mode. The former only has fields for the compile graph as the
rest is not provided. This is a small optimization but a large
majority of sync will only contain the compile Graph, as the full
dependencies are only passed during full syncs for display in the
Studio UI (Project Structure Dialog.)

Test: all tests pass.

Change-Id: I520b0ee8485594eb893a16cadaae2ec4d419a04e