Min sdk 17 for android-studio-gradle-test
We are adding a sync task that will trigger the test variant
build for the library. This will try to merge to dependency
manifests with the library one. Considering that the
min sdk is 16 for the library and for the dependency is 17,
the merging fails. The library min sdk is 17 now.

Test: ./gradlew :init performanceTest --no-daemon --info
Change-Id: Ia4dd882726f9a0217842f2b4dd74638e8b99c8a1