Moved SymbolLoader to SymbolIO.
Made a few changes to the way symbols are currently
handled:

- Added symbol class to symbol.
- Instead of directly reading from the SymbolLoader,
the SymbolWriter will read from an intermediate format,
the SymbolTable.
- Instead of using SymbolLoader to read symbols, the
code to load has been moved to SymbolIo.

This should produce no semantic changes, but should allow
for an in-memory representation of the symbol information
(the symbol table) that is independent of the way it is
loaded.

Change-Id: I2f1c1caa3d3ebb95050ce397139859ed145ad198
Test: Included.