Moved SymbolLoader to SymbolIO.
Made a few changes to the way symbols are currently
handled:

- Added symbol class to symbol.
- Instead of directly reading from the SymbolLoader,
the SymbolWriter will read from an intermediate format,
the SymbolTable.
- Instead of using SymbolLoader to read symbols, the
code to load has been moved to SymbolIo.

This should produce no semantic changes, but should allow
for an in-memory representation of the symbol information
(the symbol table) that is independent of the way it is
loaded.

Change-Id: I3ab52b70e8f7092480f558228d50a68da436984e
Test: Included.