Changing allocation tracking to be a range
Instead of allocation tracking enabled/disabled being
a single event where clients will need to reconcile
the start and stop events themselves, this changes the
accounting to be on the perfd-side. This resolves the
issue where the UI needs to guess where a start event
occurred when its display window only includes the end
event.

Change-Id: I20de4836cf95e1558a32672e1b1c28a321cc60f8
Test: changing new feature