Don't dump "ndk-build -n" output to gradle console
I recently added code to incrementally show the output from ndk-build
and CMake on the gradle console output. A bug in this change cause the
output of "ndk-build -n" to also be displayed. This isn't interesting to
the user and is too noisy so I'm changing it so that this output is no 
longer shown. STDERR from "ndk-build -n" will still be show because
it may contain errors the user needs to solve.

Test: Added an integration test that verifies that STDOUT for ndk-build -n
doesn't show in the console output that the end user sees.

Change-Id: I7b811dfe46e840d8a3293dee91e9a2bc15fe3157