Add ReadWriteThreadLock to synchronize threads in a JVM.
While we can use conventional locks to synchronize threads within the
same JVM, such locks are effective only when the class owning the
lock is loaded one time by a single class loader. If the class is loaded
multiple times by different class loaders, to the JVM, they are
different classes. The result is that only threads using the same class
loader are synchronized, whereas threads using different class loaders
are not.

ReadWriteThreadLock is useful when we want to synchronize *all* threads
in a JVM, even in the presence of multiple custom class loaders.

This class will be used by FileCache and ReadWriteProcessLock in an
upcoming commit.

Test: ReadWriteThreadLockTest
Change-Id: I83870ddbb4c0de9af1b124c75e125f20e23c11f8