Making AndroidVersionException() public
Since AndroidVersion is now loaded by the main classloader and
AndroidVersionHelper is loaded by the plugin classloader,
AndroidVersionException() can no longer be package-private.

The problem occurs if you have an old-style sdk package (with no
package.xml) of a type that is API-specific but has no API specified in
source.properties.

Test: It seems like neither the ant tests nor the bazel tests run with a
classloader setup similar to what we have in the actual app, which makes
this difficult to test...

Fixes b.android.com/226121

Change-Id: I7a654db577d0c4b35d5442cb72aab7cc34c3ec16