Http[s]URLConnection$ no longer throws "connected" exception
A user's code that called both "connect" and then "getInputStream"
would trigger an "Already Connected" exception when trying to fetch
request properties.

In other words, this would break when profiled:

> HttpURLConnection c = ...;
> c.connect();
> c.getInputStream();

Note that, in the above snippet, calling "connect" is not required, as
"getInputStream" would call it anyway as a side effect. But at the same
time, the above code is valid.

In our support library, we track both request headers and response
fields in both "connect" and "getInputStream" methods. This is
problematic in the above case because the implementation of
Http[s]URLConnection on Android throws an exception if you try to call
"requestResponseProperties" anytime after a connection has already
been made.

By the time we try tracking the response properties a second time
(in the above "c.getInptStream()" call), we are already connected and
trigger the exception.

We work around this by only tracking headers and properties once, in
either "connect" or "getInputStream", whichever is called first.

Bug: b/33666813

Test: Checking in without tests for now, as this is a P0, plus testing
supportlib code will be a very large project (and may change soon when
we make a planned switch to our approach anyway). Therefore, tests to
come later. For now, this change was manually tested by running
DisplayingBitmaps (Https) and NetworkConnect (Http) Android samples
under our profiler.

Change-Id: Ifb733354cc86f0be4000eb70ba4db1b5e5e34119